@@ -21,7 +21,7 @@ function createWindow() {
 
   mainWindow.loadURL(
     url.format({
-      pathname: path.join(__dirname, `/dist/frontend/index.html`),
+      pathname: path.join(__dirname, `/dist/frontend/browser/index.html`),
       protocol: "file:",
       slashes: true,
     })
